requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
jinja2==3.1.2
python-multipart==0.0.6
aiosqlite==0.19.0
greenlet==3.0.3
bcrypt==4.0.1
python-jose[cryptography]==3.3.0  # JWT token
passlib==1.7.4  # 密码哈希
python-jose[cryptography]==3.3.0  # JWT 认证
pydantic[email]==2.4.2  # 数据验证
gunicorn==21.2.0  # 用于生产环境的 WSGI HTTP 服务器
python-dotenv==1.0.0  # 用于环境变量管理 